Feature: alpha 설정 방식 UISlider로 변경
- StatusSection.alphaStatus를 UIStepper에서 UISlider로 바꾸었다.
  - 부드러운 투명도 변화를 위해, Model의 alpha값도 float로 바꿀 예정
- ViewController에서 View와 Model에 투명도 정보를 전달할 때, Float로 변환하여 전달하도록 함수의 매개변수를 수정

+++ DrawingApp/DrawingApp/Model/Plane.swift
@@ -29,7 +29,7 @@ struct Plane {
         rectangle[id]?.B = B
     }
 
-    func setRectangleAlpha(id: String, alpha: Double) {
+    func setRectangleAlpha(id: String, alpha: Float) {
         rectangle[id]?.alpha = Int(alpha)
     }
